Share gerrit_index.config between pods when using elasticsearch
When a new pod was added to the StatefulSet, the indices were reindexed,
which could become expensive, if the indices were large.

To avoid this, the gerrit_index.config is now shared between Gerrit pods
and reindexing will only happen, if the file does not exist or indices
are not ready.

Change-Id: Idb2ec8965e2b4c4bdd84902e9934652553db70b0