Fix versions of python packages installed with pip in containers
The gerrit-init container installs python packages with pip. The
versions of the packages were not fixed. This could potentially lead to
compatibility issues and already made it hard to know which version was
used exactly.

This change pins the version of the pymysql and the sqlalchemy packages
to the versions also already used for the tests.

Change-Id: Ic7493ac56af5f0c2aea34af3ff0a2c75486eb59e