Add annotation to Gerrit-ingresses to allow larger body size
By default the Nginx-ingress blocks large request bodies (> 1 Mb),
which resulted in a returned status of 413, when trying to push commits
containing larger amounts of data.

This change adds an annotation to the ingress for the Gerrit master that
allows to configure the maximum request body size. The size limit is
configurable in the helm chart and defaults to 50 Mb. The annotation,
which already existed in the Apache-git-http-backend component of the
gerrit-slave chart and was set to an unlimited body size, is now also
configurable.

Change-Id: Ib0d6b412f6b1a3ae4aaba9f9848043f9ae02cfc6