Allow addition of additional annotations to Ingress resources
Ingress resources can be configured by setting annotations. So far this
was abstracted away in the helm charts, but depending on the setup the
configuration might vary quite a bit and thus this abstraction did not
make sense in every case. E.g. setting an alias host name is probably
not required in a lot of use cases.

This change adds an option to the helm charts to add annotations to the
ingress resources manually and removes the abstracted implementation of
setting an alias hostname.

Change-Id: I4906ad7e4f6f64b7fd614822260f53318061e9a1