Do not enforce the usage of a ReviewDB
Gerrit does not use the ReviewDB anymore, thus no database is required.
So far the gerrit-init container was expecting a database and was
failing, if none was provided.

This change adds a configuration to explicitely enable, whether
a ReviewDB is used. This configuration is used to configure the
initContainer, whether to expect a database.

Change-Id: I0feac4b0bdeb0fada92d5f80f007393ba56bbd98