Apply file permissions to repositories on pod startup
When using the built-in Kubernetes functionality to mount NFS in
Read-Write-Many mode, the securityContext.fsGroup is neglected. This
would not allow Gerrit to write repositories into the volume, when NFS
was used.

This change provides a quick workaround by adding an initContainer that
sets the file permissions manually. However, if the volume is large this
takes a considerable amount of time. Thus, this should not be the final
solution.

Change-Id: Ib5359174d3120882a9e95adf47de69216c1ad1b8