Apply file permissions to repositories on pod startup
When using the built-in Kubernetes functionality to mount NFS in
Read-Write-Many mode, the securityContext.fsGroup is neglected. This
would not allow Gerrit to write repositories into the volume, when NFS
was used.

This change provides a workaround by adding an initContainer that sets
the file permissions manually, if the directory is not already owned by
the group used by Gerrit. This workaround is optional and can be enabled
in addition to setting the fsGroup in the securityContext, if required.

Change-Id: Ib5359174d3120882a9e95adf47de69216c1ad1b8