Add option to store TLS-certificates in external secrets
The TLS-certificates were stored in secrets managed by the helm charts.
This also required to store the certificate in the values.yaml file.
For some use cases it would be advantageous to manage the certificates
independently of helm, e.g. to not expose them in the values.yaml file
or to update them independently of helm.

This change adds an option to the gerrit-master and gerrit-slave helm
charts to give a name of an external secret containing the certificates
and to use this secret instead of creating one. For simplicity, the
default continues to be, that the secret will be created by helm.

Change-Id: I189550dea1edf737d32818a763f69a8e16ccb424