Do not enforce ReviewDB in the gerrit-slave chart
The gerrit-slave chart did expect to use a ReviewDB. Using a NoteDB was
not yet supported.

This change allows to configure, whether to use a ReviewDB. If a NoteDB
is used, a script also checks, whether the repositories containing the
NoteDB are present and waits until this is the case.

Change-Id: I824a8136e087e9287ba3b0a1c0325797092a32a0