Use alpine as a base image
So far ubuntu:18.04 was used as a base images for all container images
used in this project. Ubuntu requires quite a lot of disk space, making
the container images quite large.

This change adapts the containers to use the alpine:3.10.0 image as
a base image. Alpine is a lot smaller leading to images that are about
25-30 % of the size of the old ones.

Change-Id: Ib6aa285bdda12870f5e208aa2ea01cd22c9180cd