Package gerrit init scripts into executable
The scripts for gerrit initialization were distributed in the container
images as python source code. This required the presence of the full
dependencies in the image, making it unnecessarily large.

This change makes use of a multistage build to freeze the python source
code into an executable using pyinstaller. This reduces the image size.

Change-Id: I3a9d927d6c00e2cdb6c8f4607d3daf8bf8ba41a7