Fix slave-initialization with H2-database
The Gerrit-slave was not initialized correctly, if a password-protected
H2-database was used, since the password was not taken into account,
when initializing not running the test mode.

With this change the init-script is run with every start. This is not an
issue anymore, since the script now checks, whether an init is required.
For the initial init the volume containing the repositories will not be
symlinked to the site before init, if the test mode is disabled, to
avoid a tree in the repositories that does not fully represent the
master's tree, since the init will create the All-Projects and All-Users
repository, which would then be pushed to by replication.

Change-Id: I2df49bb4fb03a7b06fa206b41e8e94ab53edf392