Force symlink of keystore, when linking mounted components
During initialization of Gerrit, external components are mounted into
locations outside of the side and symlinked to their respective
locations in the side. For the optional keystore this symlinking was not
enforced, so that the operation failed, if the symlink already existed.

Now the symlinking of the keystore is enforced as for the other
components.

Change-Id: Ic419d6d6e8b550eb63ce30f53536397af14e5538