Add fluent-bit sidecar container to gerrit-master
So far the logs written by Gerrit were just tailed to the container's
stdout and not collected centrally.

This change adds an option to the gerrit-master chart to add
a fluent-bit sidecar container to the gerrit-master pod that collects
all logs in the /var/gerrit/logs directory and forwards it to an
ElasticSearch client (other log indexing solutions may be supported
later).

Change-Id: I3849938e1f595864704c1215b7b49f85fd6b7e91