Enable Istio as a service mesh for the Gerrit replica chart
The Gerrit replica chart could only use the Kubernetes Ingress as a
method to control incoming external traffic. Istio provides a lot
of additional control for network traffic in the cluster, which
will make it easier to build a scalable Gerrit setup.

This change adds a basic configuration for istio and expands the
chart, so that it can be configured to use the istio ingress gateway
instead of the Kubernetes Ingress.

Change-Id: Ic9d8c459c55fb0a7fd22c0ab26b9f6cec1bf2b24