Allow the use of an external PVC for git repositories
The persistent volume used to store git repositories was managed by helm.
That meant that the volume was deleted when uninstalling the chart. It
would be safer to manage the volume separately, in case an update of a
chart fails.

This change adds the option to name an external persistent volume claim
to be used for storing the repositories. That way the volume still exists,
after the chart is uninstalled and can be rebound from a new installation.

Change-Id: I3e52aa21629600771168e71c63d6e88f7a2709e8