Use healthcheck plugin for readiness/liveness-probes for Gerrit replica
The Gerrit replica pod did not yet have readiness or liveness probes.
Since the gerrit replica does not expose the REST API, this cannot be
achieved with Gerrit core. Thus, the healthcheck plugin was used
instead.

To ensure that this works, the healthcheck plugin is built into the
Gerrit container image and will be installed in any case. However, by
configuring a packaged or downloaded version in the helm chart this can
be overwritten.

In addition, a default configuration is provided to make the healthcheck
work on new Gerrit servers without existing data. To do that the
querychanges and auth healthchecks are disabled by default. This config
can be overwritten in the helm chart.

Change-Id: I15db8e86286404c7ba1f1660d4b82f0e80d9d6bc