Fix the helm chart integration tests
Helm was waiting for all PVCs to be bound, if the '--wait' flag was
set. This included the PVC for the gitGC CronJob, which would likely
not run exactly, when the integration tests were started. Thus, helm
was timing out while installing the helm charts during the integration
tests.

To avoid this, the persistence of the logs of the gitGC CronJob is
disabled for the tests.

Change-Id: Ibfc805d0a7a7bcf32b5745c285b4481e8977e5b3