Persist logs in gerrit-replica chart with multiple replica pods
When using multiple replicas of the apache-git-http-backend and/or
the gerrit-replica pods, persisting the logs did not work, since the
volumes created for this purpose were only supporting read-write-once.

A solution would have been to use StatefulSets instead of Deployments
to ensure a distinct volume for each pod, but this would give the pods
more state than required. E.g. there might be a lot of dangling volumes,
when scaled down. Other features of StatefulSets would not be used like
fixed hostnames. Thus, this solution seemed overly complex for the Gerrit
replica.

Instead now a volume providing read-write-many access, e.g. NFS, is used
to store the logs of all pods. Each pod will create a directory in the
NFS-volume with the pod name as directory name, allowing to trace the
logs to a given pod. The NFS-workaround was enabled for this additional
volume as well. Persisting logs is still optional. The persisting of the
whole Gerrit site in the gerrit-replica pods was removed. Since the site of
a replica is pretty static apart from the repositories and logs, this should
not be required and creating the site on pod creation should be fast enough.

Change-Id: I924fbd9f9cae9604bfb0af7f703aaf854a14220e