Use LEAST_CONN load balancing for gerrit-replica
Istio provides multiple load balancing algorithms. The default is
round robin. For Gerrit this is not ideal, because requests can be
quite different in size, e.g. cloning a small repo or a large repo.
Istio also provides the LEAST_CONN algorithm, which randomly selects
two live targets and selects the one with less active connections.
This would mitigate the issue and distribute the traffic better.

Change-Id: Ia12a571fdfb354dae6a264f3dbcb7a203e6e0bc6