Only use a single secret for git-backend TLS
Two separate secrets to manage the TLS certificates for the service
and ingress of the apache-git-http-backend were used. However, there
is no reason to use different certificates.

This change merges the two secrets. This should also make it easier
to use TLS-certificate forwarding from the ingress to the service.
However, this has to be configured manually, since this is different
between ingress controllers.

Change-Id: Id0e1dca75b611b2d62114b89ee5e88dc57502814