Add option for CA certificate
The download of plugins during the init could fail, if the download
server used custom certificates, e.g. in a corporate intranet.

Now a CA-certificate can be configured that will be used for TLS
verification, if the default certificates fail to validate. This
way servers with custom certificate authorities and servers with
publicly signed certificates can be used.

The provided CA will also be used by other applications in the
deployment. Right now, this is only Promtail.

Change-Id: I8ea3ef15405a939e1ce4b92c6f85f71db1e38bce