Change istio routing to remove path prefixes
The istio VirtualService was configured to use path prefixes to
distinguish the apache-git-http-backend and the gerrit-replica. This
was not required, since the `/git`and `/new` paths used by the
apache-git-http-backend are not used by Gerrit.

Now all traffic to the `/new`and `/git` paths are routed to the
apache-git-http-backend, while all other traffic is routed to the
Gerrit replic.

Change-Id: I6999c28e2791985b2b7f2bc71cadeb00da9c69a4