Enable mod_info to allow viewing configuration
In case of issues with the apache-git-http-backend it would be useful
to check the configuration that is being used at runtime. This
functionality is provided by the mod_info module.

This change activates mod_info and configures Apache to expose the
configuration on the /server-info route. This route can only be
accessed from localhost.

Change-Id: I5f25de8012cf899c399f1e7e32a79ffe6b0efac9