Use mod_cgid and multithreading instead of mod_cgi
The default configuration used preforking for multi processing.
This requires to use mod_cgi for executing CGI scripts. However,
git-http-backend allows to use threading for multi processing.
This then seems to be the better solution and can be done by
using mod_mpm_event instead of mod_mpm_prefork [1]. For
executing CGI scripts in a multi-threaded environment the
mod_cgid has to be used.

[1] https://serverfault.com/questions/383526/how-do-i-select-which-apache-mpm-to-use/383634#383634

Change-Id: Iff366444a2dcccc53c1334bb4beb625e81a97781