Set keepaliveMaxServerConnectionAge to 24h
The default configuration for the Istio pilot was to refresh the
connection to all proxies every 30m. The refreshing of the connection
and configuration required a connection reset, thereby killing all
ongoing requests. Since git requests may take quite some time, this
killed a lot of replication pushes and clones.

This mechanism is also only needed to happen often, if there the
service mesh is changing a lot by pods starting and stopping often
to rebalance the connections. This is not the case in the current
Gerrit on Kubernetes helm charts. Thus, a reconnect once per day
should be more than enough.

Change-Id: I2197d1895b2586e660e7f5e184d7f835e90c496d