Disable group indexer on startup by default
If Gerrit runs as a replica it will run a scheduled background job,
that reindexes groups to ensure that access rights to repositories
are correctly applied as on the primary instances.

By default, the first run is done offline on startup to ensure correct
group permissions in case the replica was offline for a longer time.
On large Gerrit installations, this could take several minutes. Thus,
rolling updates or pod crashes could lead to a longer time with less
active pod replicas.

In case of Kubernetes, it should be safe to turn this off by default.
Usual downtimes will be only a few minutes, e.g. during an update and
thus in the time frame of a schedule interval.
If a longer downtime was happening this option could still be enabled
once. Also a start time for the scheduler lying in the past (the default
of 00:00 is fine) should be taken, so that the indexer schedule is
directly started.

Change-Id: Ia7367aee388618f935aa621c12e59d01e0f7b71c