Set gracefulStopTimeout by default
By default the embedded Jetty server and SSHD in Gerrit will not wait for
reuqests to finish before shutting down. That means, that during
a rolling update requests will be prematurely terminated. Gerrit
however has an option to configure Jetty and Mina to wait for some time
for requests to finish, which would allow a graceful shutdown of
Gerrit and is a further step to fully zero downtime upgrades.

Change-Id: Ib646e0cf47e9219df72419f17201f65ecf71b79a