Add cronjob to regularly delete logs to gerrit-replica chart
So far, logs were kept indefinitely. This could lead to the overflowing
of the storage volumes.

This change adds a CronJob looking at the volume used to store volumes
and deleting gzipped log files older than a configured number of days.

In addition, the logging of git-gc was adapted, since the CronJob would
always store the logs in a new directory and thus the logs would never get
zipped. Now, the logs will get zipped right after the GC is finished.

Change-Id: I8cf6ce37646bc45c4110e90b9d5bac532a09854e