Merge ingresses of gerrit-replica and apache-git-http-backend
In the gerrit-replica chart the apache-git-http-backend and gerrit-
replica had each an own ingress. This is not necessary, creates
extra costs due to the need of an extra loadbalancer and is different
to the configuration in Istio.

This change merges both ingresses into a single ingress. It also removes
the TLS-endpoint of the apache-git-http-backend container, since the
TLS is anyway terminated in the ingress.

Change-Id: I526fe9653262841da8a8f5b2b655d323643ab6d9