Allow to configure additional ImagePullSecrets
With DockerHub having a rate limit now, pulling busybox could fail,
if the rate limit was exceeded. If the images provided by this project
were in a different registry than dockerhub, there was no way to
add an ImagePullSecret to the pods that allowed to log into both
registries.

This change allows to give a list of additional ImagePullSecret-names
that will be configured to be used by the pods. The secrets themselves
will have to be created manually.

Change-Id: I831a5c72c094f8366e622be13cf24a017a194a2f