Allow to configure additional secrets for Gerrit pods
Some plugins my require additional secrets to be present in some
location other than $SITE/etc.

This change adds the possibility to add such secrets. The contents
will be mounted to the Gerrit pod and the directory containing the
files of the secret will be symlinked to a configured path.

Change-Id: Ie510edc3d6ec9ed3adad38999f8ecaad75874303