Allow to configure addition ConfigMaps for Gerrit pods
Some plugins may require to persist additional configuration outside
of $SITE/etc (e.g. the saml plugin).

This change allows to configure additional ConfigMaps that load files
into a subdirectory in the $SITE/data directory. A limitation of this
change is, that the pod has to be restarted, if the ConfigMap changes.
This could be improved in the future by using a sidecar that watches
for changes in the ConfigMaps.

Change-Id: I06c63cf3e3e77d245388a095d6b81d3dd20ead35