Move adaptations of istio to profile if possible
Some adaptations made to istio could be moved to the profile. This
reduces the work required, when updating to a new version, since
they will be directly applied when recreating the manifest.

This does not work however for the addition of the lifecycle hook
in the template used for the proxy sidecar. Thus, we can't just
rely fully on the profile.

Change-Id: I829fb5726a61cc68eb6070516e66f8a80517c3bd