Allow multiple types of SSH keys
The chart so far only allowed RSA keys. However, RSA keys are not
supported anymore since OpenSSH 8.8.

This makes adding SSH keys more flexible. The way to add SSH keys
is now to use `gerrit.etc.secret` and add the SSH key files there.
The will be added to the cluster as a secret and mounted into the
Gerrit site. If non are given that way, Gerrit will create its own,
which is not recommended if horizontally scaling Gerrit due to
inconsistent host certificates.

Change-Id: If1d1a054d0d452469af2c97b7c56065ff51db851