Add k8s operator for Git GC
The plan is to use the operator pattern to manage Gerrit installations
in the future instead of relying only on helm.

This change adds a minimal operator using the Java Operator SDK.
It allows to create and delete GitGC CronJobs. This was chosen as
a starting point, since the scope is much smaller than Gerrit itself.

Change-Id: Id8f55be50c67ed27e087c9f6c00ba6aa18618db5