Remove dependency to EFS in tests
The tests were relying on AWS EFS for RWM storage. Thus, it was
impossible to run them on a different hyperscaler. Also, the used
EFS chart was deprecated a long time ago and is not compatible
with Kubernetes 1.22.

Now, the tests expect a storage class for a RWM volume to already
exist in the cluster. This way, admins can set up a suitable way to
get RWM volumes themselves according to the infrastructure being used.

Change-Id: I63c1f5a16d173825337d90bd65b06a62060f0441