Create preStop hook of istio-proxy sidecar by defaulting
To add the custom preStop hook for the istio-proxy sidecar container
allowing graceful shutdown, the istio resources were unpacked, versioned
with this repo and installed manually. However, this required to version
a lot of boilerplate code.

Istio can actually take defaults for the istio-proxy sidecar that
are added in the pod-Spec of the application and will consider
them during the sidecar injection [1].

[1] https://istio.io/latest/docs/setup/additional-setup/sidecar-injection/#customizing-injection

Change-Id: Iee5f732e4280bc2773e8a6ab246496386510f2aa