Add template for IP-whitelisting in istio
Setting up proper authentication and security is quite complicated
and will likely not always be done for development deployments.
However, if such deployments are done in the cloud, this will open
up the deployment for malicious users.

A simple way to prevent this, is to do IP-whitelisting in the Ingress.
This change adds a template to the istio profile, which allows to do
that.

Change-Id: If1556e251fbcc532737deac202832299bb818826