Allow concurrent builds
The container images could not be concurrently build on the same
host, because the base images were not tagged and the latest-tag
was used in the pipeline.

The base images are now also tagged and the latest tag is now not
used within the pipeline anymore. However, that also means that
no images tagged latest will be published.

Change-Id: I2ab80074f2e6119371f01fff247e7f8d3f19ad0e