Use --tag to overwrite default tag in ./build
Using the --tag option for ./build did not overwrite the default tag
consisting of the git describe output, but added it in addition.
If not set, it defaulted to latest. This led to issues or more
overhead, if preparing for concurrent builds, since a unique tag
always had to be set in addition to the unique tag already being the
default.

Now only the tag built from git describe is being used by default.
If --tag is used only the given tag is used.

Change-Id: I79030d43a86edae987cb422e19fbd931b9f1543c