Use LDAP auth during integration tests
Gerrit instances started during tests were not secured by authentication.
This was insecure, especially if tests were run in a public accessible
cluster.
Now LDAP-based authentication is used. A minimal LDAP server deplyoment
is provided as a supplement. The initial admin account in Gerrit is
created using selenium to log in via the Chrome browser once.

Change-Id: I07ea1592240cec7d1059016ca4d8597a78bffb32