Add CustomResource for primary Gerrit instance
With this resource users can create a Gerrit custom resource
that will be used by the operator to create the ConfigMaps,
StatefulSet and Service for a primary Gerrit.

Note that secrets will have to be provided manually and
referenced via name. The reason secrets won't be part of
the CustomResource, since they would be visible as plain
text for everybody having access to the CustomResource.

Change-Id: I9b294f317abeb7968f279773caf20c35fadc9468