Prohibit the cluster autoscaler to move running Git GC jobs
In case of a cluster scaling event, the cluster autoscaler could try
to move the pod spawned by the Git GC job to another node. This would
prematurely cancel the GC run, which in the best case would require
another GC run and in the worst case might lead to issues within the
repository.

Change-Id: I0bee0306f6c461c9f6c32cac9bc72712b8ef73d8