Copy instead of symlink cached plugins
Plugins were symlinked from the cache of downloaded plugins. This
could lead to potential issues. If a file in the cache was deleted
while still being used in another pod, using the plugin would fail.

The plugins are now copied into the Gerrit site. This also removes
the need to mount the plugin cache in the main container.

Change-Id: I530de3d7cf70ee6670b7a354f726b2ab87a82cf4