Fix default for PVC selectors in GerritCluster
The default for selectors for PVCs configured by the GerritCluster
resource was documented with `{}`. While this is valid yaml, some
provisioners don't support selectors, e.g. the nfs-subdir-provisioner.
However, the default let to an inclusion of the selector term in the
PVC manifest, which led to failing PVC creation.

Giving `null` as the default resolves that issue.

Change-Id: Ida526a2c3177c723c58f6c49ecd9b038c63a8977