[Operator] Fix updates of PVCs
PersistentVolumeClaims (PVCs) were not correctly updated. This happened
because the kubernetes client tried to set the volumeName to an empty
string. However, Kubernetes already had changed the volumeName to the
name of the provisioned PV and declares it as immutable.

This change ensures that the volumeName is not changed, if no explicit
name was set in the CustomResource.

This issue seems to be something that the client should take care of.
A solution might be to use server-side apply instead of replace in the
java-operator-sdk.

Change-Id: I75850a804eedf7ff8fc6f0badc81a6a3e9f78e36