[gerrit-replica chart] Use StatefulSet for Gerrit Replica
The Gerrit Replica was originally using a Deployment to manage the
Pods. Since the Gerrit Replica is read only and thus cares less about
state, this seemed to be the more sensible choice than a StatefulSet
as used for the primary Gerrit. A deployment is easier to handle,
e.g. contains less immutable fields, and updates faster.

However, using a Deployment required to start a job to initialize the
bare repositories before the workload pods started. This was necessary
because multiple starting pods would interfere with each other while
initializing those repositories. Further, caches and the group index
were not guaranteed to be remounted in the same pod, potentially adding
additional overhead to fill caches and reindex groups, when a pod was
restarted.

Now the Gerrit Replica uses a StatefulSet. Since pods in a StatefulSet
are started one-by-one, the job to initialize repositories is not
required anymore. As long as the same number of instances had existed
in the past, the same Gerrit site will be remounted. While caches and
indexes will still be out of date, they are at least partially filled.
This also reduces the differences between the primary Gerrit and Gerrit
Replica deployment significantly.

Change-Id: I45469d712d993be31ce12e92d5d261df1a4add85