[Operator] Enforce required configuration options
Some options like container.javaHome in the gerrit.config are defined
by the Kubernetes deployment and changing them would break the setup.

The operator now checks whether the provided gerrit.config is valid.
If a required configuration is set to an illegal value, the operator
will throw an exception and not reconcile the Gerrit deployment. If
the option is set to the expected value, the operator will do nothing.
If the option is not set,the operator will set the value.

Change-Id: I619ab47875555c5e5e10e8075880904f12b505b5