Merge gerrit and gerrit-replica container image
The only difference in the gerrit and gerrit-replica container
images was that the gerrit-replica image would fail on startup,
if the container.replica option in the gerrit.config was not set
to true. The additional maintenance effort the additional image
caused was not adequate for the gain, especially considering that
other options in the gerrit.config exist that are also not enforced
in the helm-charts but need to be set to a specific value. In the
future the operator will enforce the replica option anyway.

This change thus removes the gerrit-replica container and changes
the gerrit container to set the --replica flag on startup, if
 container.replica is set to true.

Change-Id: I14495499d057194d99c7d2fde8a2ad4f65b67d99