[Operator] Add LifecycleManager to manage synchronous shutdown hooks
In Java registered shutdown hooks are executed asynchronously. In the
case of the operator, this does not always work. If the shutdown hook
provided by the java operator SDK is executed before other shutdown
hooks, additional resources cannot be removed anymore. This however
has to be possible, if the operator should manage its own Service
and ValidationWebhooks.

Change-Id: I31dbed50c76dd3ae4e620d4a8730f606a5179789