[Operator] Add ValidationWebhook for GitGc
Currently, the operator validates whether two GitGc jobs conflict, i.e.
are working on the same project, during reconciliation. If a conflict is
detected, reconciliation was stopped and the GitGc resource marked in
the status. With this procedure the conflicting resource would however
remain in the cluster and on each restart of the operator, conflicts
would again have to be checked.

A better solution is to use validating AdmissionWebhooks. This change
adds such a webhook for GitGc resources.

Change-Id: I4329f5462757d02390157af3074625fa47b2b7ef