[Operator] Add validing webhook for Gerrit resources
The new validating webhook tests, whether a Gerrit Cluster
is referenced in the GerritSpec and ensures that only one
primary Gerrit is allowed per Gerrit cluster.

The requirement of a unique primary Gerrit per Gerrit
cluster comes from the definition of a Gerrit Cluster
and the resulting network routes. A Gerrit Cluster is
defined to work on a single set of repositories. In
case of multiple primary Gerrit's it would be impossible
to decide how to route requests, e.g. on project basis,
since not all URLs reference the projects directly.

Change-Id: I1656a2a36cda47fb3b72e7196565ffd7c09313b4