Check whether reindexing is required independently of init
The check whether reindexing was required was only done, if the
`gerrit.war init` command was executed. In cases where the init
failed at a late stage, this could prevent Gerrit to start,
since the init under some conditions would not need to rerun, but
the index was not ready.

Now the check is executed on every startup for primary Gerrits.

Change-Id: I7aa56b059fba272fd41433d6b5a9eabf8de5dab0