[operator] Check gerrit.config with Validation Webhook
If the gerrit.config contained illegal configuration, i.e. configuration
that would not work with the Kubernetes setup, the operator would not
reconcile the Gerrit resource, but the resource itself would be present
in the cluster.

Now, a validation webhook checks the Gerrit configuration. This means
that the resource can't be applied in the first place and the user gets
an early error.

Change-Id: I8f8a76866f4a36aa4e1894e382f859aebed39893