Automatically route HTTP requests between primary and replicas
If Istio is being used, the operator will now configure Istio to
automatically route fetch/clone requests to the replica and all
other requests to the primary. This way only a single hostname:port
combination is required.

This is not possible for SSH, since it is load balanced on network
layer 4, where the request content can't be accessed. Thus, for SSH
still different subdomains have to be used to directly access the
target Gerrit instance.

This feature also does not exist if using Ingresses, since Ingresses
to not support regex URI matching, but only prefix or exact, which
are not suitable to filter for fetches and clones. For some ingress
controllers this could be done via annotations, but since there are
several, it would add quite a bit of code. It should only be added,
if the use case arises.

Change-Id: I79e7ce2d09c3345a1a04d35b0925f62542a64f15