[Operator] Separate concerns between Gerrit and GerritCluster
The GerritCluster and Gerrit CustomResources had bidirectional
dependencies on each other. This caused a lot of requests to the
Kubernetes API server to find the respective other resources, since
information was required from them.

This change redesigns the way that the operator is designed. Thus,
this is a breaking change!

The GerritCluster resource now does not only deploy the storage
and network components of a Gerrit cluster, but also the Gerrit
instances that are part of the cluster. Thus, deploying Gerrit
CustomResources manually is not required anymore. In fact, it is
not possible anymore to add Gerrit resources to a GerritCluster
other than by adding it to the GerritCluster CustomResource.

A Gerrit CustomResource can still be applied to the cluster, but
it won't create the storage components, e.g. PVCs, and apart from
a service, no networking components will be installed. Thus, it is
highly recommended to use the GerritCluster custom resource instead
even if only a single Gerrit deployment will be used.

Change-Id: I69f2f6633c9d964412317c09c0dfe817b2ec0cb3