[Operator] Integrate  Receiver into GerritCluster
In a recent change the Gerrit CustomResource was integrated into
the GerritCluster CustomResource, since there were bidirectional
dependencies. The same is true for the Receiver CustomResource.

Thus, a Receiver can now directly be configured in a GerritCluster.
This is now also the only way to add a Receiver to a GerritCluster.
A Receiver can still be deployed separately, but Storage and Network
have to be created manually in that case.

Change-Id: I04d3b25d0bc5dc92e2ec6f934db3fe7e4f73ed41