Make secret used by Gerrit a read-only dependent resource
The operator does not create or modify the secret used by Gerrit.
Instead it has to be managed manually and is only referenced by name.
So far updates were detected by a manually added EventListener.

Now, a read-only dependent resource is used to get updates on the
Secret. This more naturally fits into the workflow and allows to
reference it from the reconciler context.

Change-Id: Ic89aeda3e40585a611ed87b737e888552b5d5e7f