Trigger restart of Gerrit pods during StatefulSet reconciliation
The operator triggered a rolling restart of Gerrit, when its configuration
changed. This was done after reconciliation of all dependent resources
including the StatefulSet. This was not ideal, since this caused
additional reconciliation runs.

Now, restarts are triggered during reconciliation of the dependent
StatefulSet by setting the `kubectl.kubernetes.io/restartedAt`
annotation.

This change further improves the way the necessity of the restart is
computed. This is enabled and required to trigger it at this stage.

Change-Id: I374c67e1d2a498a60c07237c4bc2748a043f7efb