Add gateway label for CI builds
This change adds a label that will be used by the CI to decide
whether it starts verifying a change. The label can only be voted
on by maintainers of the project.

This is a security feature. Verifying changes requires to deploy
the application to a Kubernetes cluster. A malicious change could
deploy some unwanted application into the cluster, e.g. to mine
bitcoins. Thus, maintainers should assess whether a change is safe
to deploy for testing.

Change-Id: Ifdd17349f6b60ab0c8faf3d817ebb3f610be90de