[Operator] Fix flaky tests
In some tests pods were updated. In that case the check whether pods
are ready could prematurely return true.

This change adds a 15s delay in polling for pod readiness to ensure
that the controllers had enough time to trigger the pod update.

It further fixes the test that should check that Gerrit pods are not
restarted if the configuration was not updated. This test updated
the StatefulSet, which triggered a rolling update independently.
However, more interesting is the case when a dependent resource that
is neither the StatefulSet nor any configuration is updated.

Change-Id: I44b9d252df3a14807e7e536fbf9c99b018c74c0e