Add helm charts for operator deployment.
Also update the documentation to reflect option to deploy the
operator via the helm charts.

The helm charts make for more convenient management of the operator
resources. Moreover, because the CRDs are in their own helm chart, the
Gerrit operator user (not developer) can skip the maven build step for
generating the CRDs (although currently they would still need to run
`maven clean install -P publish` to publish the Docker image for the
operator).

Change-Id: Ifb2eb765c3418822ec49eaec2ab4282bdf839874