[Operator] As part of build copy updated CRDs to helm-chart
The gerrit-operator-crds helm-chart had to be updated manually.
This is an error-prone process, since changes to the CRDs might
not be applied to the helm-chart.

Now, the operator build will copy the CRDs to the helm-chart,
overwriting the existing files. When creating the commit, the
updated CRDs just have to be added.

In the future, a CI should validate that the CRDs are up-to-date.

Change-Id: I71ae592cbc721981e4fe409c44bce783866eba17